services/tasks_api/pyproject.toml
[tool.poetry]
name = "tasks-api"
version = "0.1.0"
description = ""
authors = ["Joris Jansen <joris97jansen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
httpx = "^0.26.0"
uvloop = "^0.19.0"
httptools = "^0.6.1"
mangum = "^0.17.0"
boto3 = "^1.34.41"
moto = "4.2.14"
pyjwt = "^2.8.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
bandit = "^1.7.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
